samples/python/requirements.txt
azure-ai-projects>=1.0.0
azure-ai-agents>=1.1.0
azure-ai-evaluation>=1.0.0b5
azure-identity>=1.17.0
requests>=2.32.0
